TestWikipedia
My test task for android junior.

Make a client server application.

The first screen contains an input field and a button Search.

By clicking the button application validates the input field if it is
not empty, a made request to the server Wikipedia, otherwise, the
message "Input field is empty. "

This result is parsed (json) and displayed on a new screenshot in a
list.

Request to the server by service in a separate thread. The fact that the
query executed report through the Broadcast.